Refactored `filetypes/__init__.py`
I renamed PixeFile to Factory, because that's what it is. I also moved it out of `__init__.py` and placed it in `base.py` to keep large chunks of code out of `__init__.py`.

I also created a PixeFile parent class to use as a default for a bunch of stuff and to give a basic structure for additional filetype classes.